requirements.txt
opencv-python
numpy 
pandas
matplotlib
flask
pyserial